refactor(bm13xx): split protocol into submodules
The protocol module held register definitions and payloads,
command framing, response decoding, and the serial codec in one
file. Move each into its own sibling module.
