+++ mujina-miner/src/asic/bm13xx/mod.rs
@@ -4,9 +4,13 @@
 //! communicating with BM13xx series mining chips (BM1366, BM1370, etc).
 
 pub mod chip_config;
+pub mod codec;
+pub mod command;
 pub mod crc;
 pub mod error;
 pub mod protocol;
+pub mod register;
+pub mod response;
 pub mod thread;
 
 #[cfg(test)]
@@ -15,7 +19,9 @@ mod reference_tests;
 pub mod test_data;
 
 // Re-export commonly used types
-pub use protocol::{FrameCodec, Register, Response};
+pub use codec::FrameCodec;
+pub use register::Register;
+pub use response::Response;
 
 // Re-export the protocol handler
 pub use protocol::BM13xxProtocol;
